Add: acceptSslCerts config option
This is pretty important for testing locally using a self-signed cert.

I have defaulted it to false just as BrowserStack does.

I appreciate your fork is not the original, but it makes sense for you to merge
this change with your pull request for the full BrowserStack functionality
rather than me wait for it to be merged then wait again for Mink to merge this
option.