Fixed configuration for Selenium2 factory
Changing the scalarNode to a booleanNode actually fixes a bug in the latest version of IEDriverServer + Selenium2, which triggers a segmentation fault if the value given for ignoreZoomSetting is a string and not a boolean (which is the case currently).
A simple check for this is to try launching Behat + mink + Selenium2 with Selenium2 + IEDriverServer, with the following defaults:

```yml
default:
  suites:
    default:
      contexts:
        - Behat\MinkExtension\Context\MinkContext
  extensions:
    Behat\MinkExtension:
      javascript_session: default
        sessions:
          default:
            selenium2:
              browser: "internet explorer"
```

Adding the capabilities with ignoreZoomSetting set to true or false also fixes the issue, but not at the source of the problem.