Add some doc of the @insulated tag
It is a bit difficult to know how much detail to write. For example, Mink ``resetSessions()`` just does ``deleteAllCookies()`` - but writing that detail here would mean that it depends on the implementation of Mink (which might do different things in future), and so such doc here would go out of date and be misleading.

The reason I found this stuff is that Microsoft Edge has a problem with its webdriver and deleteAllCookies(). The workaround for the Behat-MinkExtension-Mink-Selenium-EdgeWebDriver chain is to use ``@insulated`` on all scenarios. Then the code path never calls ``deleteAllCookies()`` so we get lucky and can do Edge UI testing with Behat... Can't/shouldn't really document that here! ``@insulated`` was not intentionally provided as a fix for the Edge webdriver!